requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
docopt==0.6.2
idna==3.2
loguru==0.5.3
pip-upgrade==0.0.6
pipreqs==0.4.10
pyTelegramBotAPI==4.1.0
python-decouple==3.5
python-dotenv==0.19.1
requests==2.25.1
telebot==0.0.4
urllib3==1.26.7
win32-setctime==1.0.3
yarg==0.1.9
pip==21.2.4
wheel==0.36.2
setuptools==58.1.0